Split tests into CLI + Create
